test: separate retrieval api contract integration coverage
