+++ docs/architecture/system-overview.md
@@ -100,6 +100,9 @@ Prompt API behavior is also verified through its own dedicated integration modul
 governance and runtime contract coverage can evolve independently of the broader platform
 health suite.
 
+Retrieval API behavior is also verified through its own dedicated integration module so
+retrieval contract coverage can evolve independently of the broader platform health suite.
+
 ### Async Runtime
 
 - `src/app/services/async_runtime_status.py`
